✨ (Menu) - Add a menu on the top of the app
