+++ src/app/layout.tsx
@@ -5,6 +5,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import { ThemeProvider } from "@/components/theme/ThemeProvider";
+import Menu from "@/components/Menu";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -35,6 +36,7 @@ export default function RootLayout({
           <div className="fixed bottom-4 right-4 z-50">
             <ThemeToggle />
           </div>
+          <Menu/>
           {children}
           <Toaster />
         </ThemeProvider>
